7175 Want improved 6513 detection and handling
This patch introduces a feature flag, hole_birth_fix, whose purpose is to
provide a txg beyond which we are certain fixes for the bugs in 6513 and 7176
have been merged. It also changes the default behavior for pools where this flag
is not enabled and txgs prior to the flag being enabled, to not trust the
birth txg metadata on holes (due to 6513 and 7176).