Drop apparently-unneeded byteswap check of c_len.
I have a dataset which was generated on what I thought was a
Linux/sparc64 system unmodified except for fixing the missing
__bswap[sd]i2 functions, in which the c_len is incorrectly
byteswapped.

But I cannot reproduce this using a copy of git master modified
in the way described above, so I am forced to conclude I must
have mistakenly generated it using one of my experimental
versions while I was playing with the byteswap usage.

So I'm dropping this adaptation.

Signed-off-by: Rich Ercolani <rincebrain@gmail.com>