Avoid accessing BRT on every block free.
Currently the biggest concern with BRT is the need to check BRT table
on every data block free - if the block is there we need to decrement
counter, if it is not in there we simply free the block. In majority
of cases the block won't be in BRT. If BRT won't fit into memory we
will have the same problems as with deduplication.

To address that create an array of reference counters for every vdev
(vdev array). A single counter represents 1GB of vdev storage.
Keep vdev arrays in memory: for 1PB vdev we need 8MB of memory.

If a block is being referenced through BRT we also bump reference
counter for the 1GB chunk of where the blocks starts. If we have a
data block at offset 0 which is referenced 3 times, another data
block at offset 131072 which is referenced 2 times and one more
data block at offset 262144 which is referenced 4 times, we will
have 9 (3+2+4) in the vdev array at index 0.

When we sync transactions group and update BRT entries we also sync
vdev arrays if needed, but the vdev arrays are never removed from memory.

When we are about to free a block we check the vdev array. If the
array element at index (blk_offset/1GB) is greater than 0 then it is
possible that we may have a BRT entry for this block, so we look it up.
If it is 0 then we are sure there is no BRT entry and we can free
the block.

For very large vdevs, vdev array might potentially span multiple blocks
on disk. If we update just a single entry in vdev array it would be
a waste to write entire array to the disk. It is not a huge deal -
vdev array for 8TB vdev is 64kB, but still... To address that maintain
in-memory bitmap where a single bit represents as many array elements as
we can fit in a (1 << ashift) block. Any change to the vdev array will
set a bit in the bitmap, so we know that this specific block of the vdev
array is dirty and needs to be synced to the disk.
This is partially implemented.

Adding new vdevs is supported, so is growing vdevs.
Removing vdevs and shrinking them is not yet supported.