Connect BRT to the ZIO pipeline.
We only need additional stage for BRT free.

Note that if the block is in both BRT and DDT we can choose from where
to free it first by deciding about zio_ddt_free() vs zio_brt_free()
order. The current implementation executes DDT free first, so the
block is first removed from DDT even if it was added there later.

This can also be used as a hack to reduce DDT size - we clone files
using BRT and remove then, which will free the blocks in DDT, but
leave them in BRT.