Make it rain SIMD on FreeBSD
This is a conflation of what will probably end up being separate
patches - enabling SIMD on ARM64 FreeBSD, doing so for at least
fletcher4, and enabling a version of superscalar4 fletcher4 built
with the compiler's own notions of SIMD optimization after noticing
that was a huge win on some ARM platforms.

INCOMPLETE: This patch hasn't been configured to build on Linux
properly, the raidz support isn't wired up much less tested, I want
to do much more testing on ARM FBSD...

(The init/fini functions are handled specially because at least
on FBSD/ARM64, the compiler tries to implement the bzero/memset call
as load into a vector register, so we'd need to add a whole save/restore
just for that or refactor the code to know when to do it outside of the
specific implementations, and that seemed like a fair bit of work for
not an obvious win...)

Based on an initial patch by Allan Jude.

Signed-off-by: Rich Ercolani <rincebrain@gmail.com>