Force scalar fletcher4 functions to remain scalar
UBSan warned us that we were telling the compiler this was
required to be 64B aligned, and then not doing it, and we
ignored it.

In the kernel, this didn't matter, because the scalar accesses
were forced to not be vectorized and the alignment wasn't required
by anything else.

In userland, if you enable auto-vectorization (which gcc 12 now
does at -O2), the compiler will believe you when you promised it
that ctx->scalar would be 64B aligned, and generate aligned write
instructions...and then if it is not, in fact, aligned, you crash.

Signed-off-by: Rich Ercolani <rincebrain@gmail.com>