Lean into it
Okay, so just forcing no vectorization fixes the crash, but is still
UB.

It turns out, almost everywhere who's calling fletcher_4_scalar_native
is just coercing a zio_cksum_t * into a fletcher_4_ctx_t *, so if
that's ever not valid, we're going to crash and burn hard. So let's
just do that in the actual scalar functions, to avoid the compiler
thinking it's really an aligned type when it's not.

Signed-off-by: Rich Ercolani <rincebrain@gmail.com>